extracted components to separate files, connected SettingsContainer to store
