+++ applications/launchpad_v2/src/components/Link/index.tsx
@@ -1,0 +1,9 @@
+import { AnchorHTMLAttributes } from 'react'
+
+import { StyledAnchor } from './styles'
+
+const Link = (props: AnchorHTMLAttributes<HTMLAnchorElement>) => (
+  <StyledAnchor target='_blank' {...props} />
+)
+
+export default Link
