+++ applications/launchpad_v2/src/components/Link/styles.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const StyledAnchor = styled.a`
+  color: ${({ theme }) => theme.accent};
+  text-decoration: none;
+  &:hover {
+    color: ${({ theme }) => theme.accentDark};
+  }
+`
