+++ applications/launchpad_v2/src/components/TitleBar/types.ts
@@ -1,4 +1,3 @@
 export interface TitleBarProps {
   drawerViewWidth?: string
-  openSettings: () => void
 }
