+++ applications/launchpad_v2/src/containers/SettingsContainer/types.ts
@@ -1,0 +1,12 @@
+import { Settings } from '../../store/settings/types'
+
+export type SettingsProps = {
+  onSettingsTouched: (changed: boolean) => void
+}
+
+export type SettingsComponentProps = {
+  open?: boolean
+  onClose: () => void
+  goToSettings: (s: Settings) => void
+  activeSettings: Settings
+}
