+++ applications/launchpad_v2/src/layouts/MainLayout/index.tsx
@@ -29,8 +29,6 @@ const MainLayout = ({ drawerViewWidth = '50%' }: MainLayoutProps) => {
   // Decrease the padding when the main container becomes 'small'
   const [tightSpace, setTightSpace] = useState(false)
 
-  const [settingsOpen, setSettingsOpen] = useState(true)
-
   const [expertViewSize, invertedExpertViewSize] =
     ExpertViewUtils.convertExpertViewModeToValue(expertView, drawerViewWidth)
 
@@ -71,10 +69,7 @@ const MainLayout = ({ drawerViewWidth = '50%' }: MainLayoutProps) => {
 
   return (
     <ScreenContainer>
-      <TitleBar
-        drawerViewWidth={drawerViewWidth}
-        openSettings={() => setSettingsOpen(true)}
-      />
+      <TitleBar drawerViewWidth={drawerViewWidth} />
 
       <MainLayoutContainer>
         <MainContainer
@@ -104,10 +99,7 @@ const MainLayout = ({ drawerViewWidth = '50%' }: MainLayoutProps) => {
           <ExpertView />
         </ExpertViewDrawer>
       </MainLayoutContainer>
-      <SettingsContainer
-        open={settingsOpen}
-        onClose={() => setSettingsOpen(false)}
-      />
+      <SettingsContainer />
     </ScreenContainer>
   )
 }
