+++ applications/launchpad_v2/src/store/index.ts
@@ -3,12 +3,14 @@ import { configureStore } from '@reduxjs/toolkit'
 import baseNodeReducer from './baseNode'
 import walletReducer from './wallet'
 import appReducer from './app'
+import settingsReducer from './settings'
 
 // exported for tests
 export const rootReducer = {
   app: appReducer,
   baseNode: baseNodeReducer,
   wallet: walletReducer,
+  settings: settingsReducer,
 }
 
 export const store = configureStore({
