+++ applications/launchpad_v2/src/store/settings/selectors.ts
@@ -1,0 +1,4 @@
+import { RootState } from '../'
+
+export const selectSettingsOpen = (state: RootState) => state.settings.open
+export const selectActiveSettings = (state: RootState) => state.settings.which
